Refactored API, utils and workers. Added more commends to funtions.
