+++ update_tracks_worker.py
@@ -1,12 +1,12 @@
-from app.API.spotify import get_access_token, StatusCodeError
-from app.tasks import update_user_tracks
-from app.models import User
+from app.utils.spotify import get_access_token, StatusCodeError
+from app.utils.tasks import update_user_tracks
+from app.utils.models import User
 import requests
 import sys
 
 if __name__ == '__main__':
     # We Limit the traceback to keep the log files clear.
-    # sys.tracebacklimit = 0
+    sys.tracebacklimit = 0
 
     # Update user tracks
     refresh_tokens = User.get_all_tokes()
